requirements.txt
azure-mgmt-resource>=18.0.0
azure-mgmt-web
azure-identity>=1.5.0
azure-mgmt-rdbms
azure-mgmt-storage
azure-mgmt-servicebus==6.0.0
azure-mgmt-cosmosdb
azure-keyvault-secrets
azure-mgmt-loganalytics
azure-mgmt-applicationinsights
azure-mgmt-monitor
ansible